# docs/swagger.yaml
basePath: /
definitions:
  models.AuthResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.Booking:
    properties:
      created_at:
        type: string
      id:
        type: integer
      passenger_id:
        type: integer
      requested_seats:
        type: integer
      status:
        type: string
      trip_offer_id:
        type: integer
      updated_at:
        type: string
      warning_acknowledged:
        type: boolean
    type: object
  models.CleanupExpiredTripsResponse:
    properties:
      expired_trips_count:
        type: integer
      message:
        type: string
    type: object
  models.CreateBookingRequest:
    properties:
      requested_seats:
        minimum: 1
        type: integer
      trip_offer_id:
        type: integer
      warning_acknowledged:
        type: boolean
    required:
    - requested_seats
    - trip_offer_id
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.SendOTPRequest:
    properties:
      phone_number:
        type: string
    required:
    - phone_number
    type: object
  models.SendOTPResponse:
    properties:
      expires_at:
        type: string
      message:
        type: string
      otp_code:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    type: object
  models.UserRole:
    enum:
    - driver
    - passenger
    - both
    type: string
    x-enum-varnames:
    - UserRoleDriver
    - UserRolePassenger
    - UserRoleBoth
...
  models.VerifyOTPRequest:
    properties:
      code:
        type: string
      phone_number:
        type: string
    required:
    - code
    - phone_number
    type: object
info:
  contact: {}
  description: MVP backend for Safar ride-sharing platform
  title: Safar Backend API
  version: "1.0"
paths:
  /admin/cleanup/expired-trips:
    post:
      description: Marks trips older than 7 days after trip_date as expired.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CleanupExpiredTripsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cleanup expired trips
      tags:
      - Admin
  /auth/send-otp:
    post:
      consumes:
      - application/json
      description: Generates and stores OTP code for a phone number.
      parameters:
      - description: Send OTP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SendOTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Send OTP code
      tags:
      - Auth
  /auth/verify-otp:
    post:
      consumes:
      - application/json
      description: Verifies OTP and returns JWT access token with user profile.
      parameters:
      - description: Verify OTP request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Verify OTP code
      tags:
      - Auth
  /bookings:
    post:
      consumes:
      - application/json
      description: Creates a passenger booking for an active trip if seats are available.
      parameters:
      - description: Create booking request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - Bookings
  /bookings/{id}/cancel:
    patch:
      description: Cancels a passenger booking and restores seats to the trip.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel booking
      tags:
      - Bookings
  /bookings/my:
    get:
      description: Returns bookings owned by the authenticated passenger.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my bookings
      tags:
      - Bookings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"